Supervised/Linear regression/linear_regression_project/requirements.txt
pandas
numpy
scikit-learn
matplotlib
jupyter
PyYAML